+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Transient subvolumes no longer create orphaned snapshots when no drives can receive sends
+- Transient subvolumes no longer create snapshots when send interval hasn't elapsed
+
 ## [0.12.0] - 2026-04-05
 
 ### Added